requirements.txt
yt-dlp>=2023.3.4
youtube-transcript-api>=0.6.1
python-dotenv>=1.0.0
google-generativeai>=0.7.0